refactor: `NetworkManagerIndicator` component to separate file.
